Move new/ page components to components directory
